+++ crates/ruma-events/src/call/member.rs
@@ -269,14 +269,15 @@ mod tests {
     use crate::{
         AnyStateEvent, StateEvent,
         call::member::{EmptyMembershipData, FocusSelection, SessionMembershipData},
+        rtc::notification::CallIntent,
     };
 
     fn create_call_member_legacy_event_content() -> CallMemberEventContent {
         CallMemberEventContent::new_legacy(vec![LegacyMembershipData {
-            application: Application::Call(CallApplicationContent {
-                call_id: "123456".to_owned(),
-                scope: CallScope::Room,
-            }),
+            application: Application::Call(CallApplicationContent::new(
+                "123456".to_owned(),
+                CallScope::Room,
+            )),
             device_id: owned_device_id!("ABCDE"),
             expires: Duration::from_secs(3600),
             foci_active: vec![Focus::Livekit(LivekitFocus {
@@ -290,10 +291,7 @@ mod tests {
 
     fn create_call_member_event_content() -> CallMemberEventContent {
         CallMemberEventContent::new(
-            Application::Call(CallApplicationContent {
-                call_id: "123456".to_owned(),
-                scope: CallScope::Room,
-            }),
+            Application::Call(CallApplicationContent::new("123456".to_owned(), CallScope::Room)),
             owned_device_id!("ABCDE"),
             ActiveFocus::Livekit(ActiveLivekitFocus {
                 focus_selection: FocusSelection::OldestMembership,
@@ -372,10 +370,7 @@ mod tests {
     #[test]
     fn deserialize_call_member_event_content() {
         let call_member_ev = CallMemberEventContent::new(
-            Application::Call(CallApplicationContent {
-                call_id: "123456".to_owned(),
-                scope: CallScope::Room,
-            }),
+            Application::Call(CallApplicationContent::new("123456".to_owned(), CallScope::Room)),
             owned_device_id!("THIS_DEVICE"),
             ActiveFocus::Livekit(ActiveLivekitFocus {
                 focus_selection: FocusSelection::OldestMembership,
@@ -421,13 +416,114 @@ mod tests {
     }
 
     #[test]
+    #[cfg(feature = "unstable-msc4075")]
+    fn deserialize_event_with_call_intent() {
+        let call_member_ev = CallMemberEventContent::new(
+            Application::Call(CallApplicationContent {
+                call_id: "".to_owned(),
+                scope: CallScope::Room,
+                call_intent: Some(CallIntent::Audio),
+            }),
+            owned_device_id!("THIS_DEVICE"),
+            ActiveFocus::Livekit(ActiveLivekitFocus {
+                focus_selection: FocusSelection::OldestMembership,
+            }),
+            vec![Focus::Livekit(LivekitFocus {
+                alias: "room1".to_owned(),
+                service_url: "https://livekit1.com".to_owned(),
+            })],
+            None,
+            None,
+        );
+
+        let json = json!({
+              "application": "m.call",
+              "call_id": "",
+              "scope": "m.room",
+              "m.call.intent": "audio",
+              "device_id": "THIS_DEVICE",
+              "foci_preferred": [
+                {
+                  "type": "livekit",
+                  "livekit_alias": "room1",
+                  "livekit_service_url": "https://livekit1.com"
+                }
+              ],
+              "focus_active": {
+                "type": "livekit",
+                "focus_selection": "oldest_membership"
+              },
+              "expires": 14_400_000
+        });
+
+        let ev_content: CallMemberEventContent = serde_json::from_value(json).unwrap();
+        assert_eq!(
+            serde_json::to_string(&ev_content).unwrap(),
+            serde_json::to_string(&call_member_ev).unwrap()
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "unstable-msc4075")]
+    fn deserialize_application() {
+        let test_cases = vec![
+            (
+                Application::Call(CallApplicationContent {
+                    call_id: "".to_owned(),
+                    scope: CallScope::Room,
+                    call_intent: None,
+                }),
+                json!({
+                  "application": "m.call",
+                  "call_id": "",
+                  "scope": "m.room",
+                }),
+            ),
+            (
+                Application::Call(CallApplicationContent {
+                    call_id: "".to_owned(),
+                    scope: CallScope::Room,
+                    call_intent: Some(CallIntent::Audio),
+                }),
+                json!({
+                  "application": "m.call",
+                  "call_id": "",
+                  "scope": "m.room",
+                  "m.call.intent": "audio"
+                }),
+            ),
+            (
+                Application::Call(CallApplicationContent {
+                    call_id: "xxxx".to_owned(),
+                    scope: CallScope::User,
+                    call_intent: Some(CallIntent::Video),
+                }),
+                json!({
+                  "application": "m.call",
+                  "call_id": "xxxx",
+                  "scope": "m.user",
+                  "m.call.intent": "video"
+                }),
+            ),
+        ];
+
+        for (model, jon) in test_cases {
+            let app: Application = serde_json::from_value(jon).unwrap();
+            assert_eq!(
+                serde_json::to_string(&app).unwrap(),
+                serde_json::to_string(&model).unwrap()
+            );
+        }
+    }
+
+    #[test]
     fn deserialize_legacy_call_member_event_content() {
         let call_member_ev = CallMemberEventContent::new_legacy(vec![
             LegacyMembershipData {
-                application: Application::Call(CallApplicationContent {
-                    call_id: "123456".to_owned(),
-                    scope: CallScope::Room,
-                }),
+                application: Application::Call(CallApplicationContent::new(
+                    "123456".to_owned(),
+                    CallScope::Room,
+                )),
                 device_id: owned_device_id!("THIS_DEVICE"),
                 expires: Duration::from_secs(3600),
                 foci_active: vec![Focus::Livekit(LivekitFocus {
@@ -438,10 +534,10 @@ mod tests {
                 created_ts: None,
             },
             LegacyMembershipData {
-                application: Application::Call(CallApplicationContent {
-                    call_id: "".to_owned(),
-                    scope: CallScope::Room,
-                }),
+                application: Application::Call(CallApplicationContent::new(
+                    "".to_owned(),
+                    CallScope::Room,
+                )),
                 device_id: owned_device_id!("OTHER_DEVICE"),
                 expires: Duration::from_secs(3600),
                 foci_active: vec![Focus::Livekit(LivekitFocus {
@@ -547,10 +643,10 @@ mod tests {
         assert_eq!(member_event.room_id, room_id);
         assert_eq!(member_event.origin_server_ts, TS(js_int::UInt::new(111).unwrap()));
         let membership = SessionMembershipData {
-            application: Application::Call(CallApplicationContent {
-                call_id: "".to_owned(),
-                scope: CallScope::Room,
-            }),
+            application: Application::Call(CallApplicationContent::new(
+                "".to_owned(),
+                CallScope::Room,
+            )),
             device_id: owned_device_id!("THIS_DEVICE"),
             foci_preferred: [Focus::Livekit(LivekitFocus {
                 alias: "room1".to_owned(),